More consistent code styling + readability refactorings.
